캐릭터 배치시 시너지 증가와 중복 제거
와 중복제거하기 진짜 힘들었어요 ㅠㅠ
무려 int 배열 두개랑 bool 배열 하나로 구현하긴 했는데
성능은 모르겠음 ㅋㅋㅋㅋ

+++ TeamFightTatics/Assets/Persona.cs
@@ -21,8 +21,8 @@ public class Persona : Character
         get { return _defaultMp; }
         set { _defaultMp = value; }
     }
-    private int _attackRange;
-    public int AttackRange
+    private float _attackRange;
+    public float AttackRange
     {
         get { return _attackRange; }
         set { _attackRange = value; }
@@ -39,7 +39,6 @@ public class Persona : Character
         get { return _attackDelayTime; }
         set { _attackDelayTime = value; }
     }
-
     [SerializeField]
     private int _diposedIndex;
     public int DiposedIndex
@@ -47,7 +46,6 @@ public class Persona : Character
         get { return _diposedIndex; }
         set { _diposedIndex = value; }
     }
-
     private void Update()
     {
         //return to initial state
